Introduce userSession object to store the logged in user. Adapt the DatabaseAPI to check user permissions before editing user data. Show edit buttons in front end only if user has edit rights

+++ samj/src/main/java/com/samj/shared/DatabaseAPI.java
@@ -3,6 +3,7 @@ package com.samj.shared;
 import com.samj.backend.CallForwardingRecordsDAO;
 import com.samj.backend.SettingsDAO;
 import com.samj.backend.UserDAO;
+import com.samj.frontend.UserSession;
 
 import java.time.LocalDateTime;
 import java.util.Set;
@@ -14,8 +15,12 @@ import java.util.Set;
  */
 public class DatabaseAPI {
 
-    public static boolean createNewUser(UserDTO userDTO) {
-        if (!Utils.validateUserDTO(userDTO)) {
+    public static boolean createNewUser(UserSession userSession, UserDTO userDTO) {
+        if (! _isUserHasEditPermission(userSession)) {
+            return false;
+        }
+
+        if (! Utils.validateUserDTO(userDTO)) {
             return false;
         }
 
@@ -25,14 +30,18 @@ public class DatabaseAPI {
 
     /**
      * In some cases, we already do the validation in the frontEnd, this method will
-     * create the new user without validating the data.
+     * create the new user without validating the data, only role validation done.
      */
-    public static boolean createNewUserWithoutValidation(UserDTO userDTO) {
+    public static boolean createNewUserWithoutDataValidation(UserSession userSession, UserDTO userDTO) {
+        if (! _isUserHasEditPermission(userSession) || ! _isUserEditingHisOwnData(userSession, userDTO)) {
+            return false;
+        }
+
         Utils.encryptUserPassword(userDTO);
         return UserDAO.createUser(userDTO);
     }
 
-    //User API
+    //UserSession API
     public static Set<UserDTO> loadAllUsers() {
         return UserDAO.loadAllUsers();
     }
@@ -49,37 +58,68 @@ public class DatabaseAPI {
         return UserDAO.loadUserByUsername(username);
     }
 
-    public static boolean deactivateUser(String username) {
+    public static boolean deactivateUser(UserSession userSession, String username) {
+        if (! _isUserHasEditPermission(userSession)) {
+            return false;
+        }
+
         return UserDAO.updateUserStatus(username, "inactive");
     }
 
-    public static boolean reactivateUser(String username) {
+    public static boolean reactivateUser(UserSession userSession, String username) {
+        if (! _isUserHasEditPermission(userSession)) {
+            return false;
+        }
+
         return UserDAO.updateUserStatus(username, "activate");
     }
 
-    public static boolean deleteUser(String username) {
+    public static boolean deleteUser(UserSession userSession, String username) {
+        if (! _isUserHasEditPermission(userSession)) {
+            return false;
+        }
+
         return UserDAO.deleteUser(username);
     }
 
-    public static boolean markUserAsDeleted(String username) {
+    public static boolean markUserAsDeleted(UserSession userSession, String username) {
+        if (! _isUserHasEditPermission(userSession)) {
+            return false;
+        }
+
         return UserDAO.markUserAsDeleted(username);
     }
 
-    public static boolean updateUserPassword(String username, String password) {
-        return UserDAO.updateUserPassword(username, password);
+    public static boolean updateUserPassword(UserSession userSession, String username, String password) {
+        if (_isUserHasEditPermission(userSession) || _isUserEditingHisOwnData(userSession, username)) {
+            return UserDAO.updateUserPassword(username, password);
+        }
+
+        return false;
     }
 
-    public static boolean updateUserAllFields(UserDTO userDTO) {
+    public static boolean updateUserAllFields(UserSession userSession, UserDTO userDTO) {
+        if (! _isUserHasEditPermission(userSession)) {
+            return false;
+        }
+
         return UserDAO.updateUserAllFields(userDTO);
     }
 
     /**
      * In some cases, we already do the validation in the frontEnd, this method will
-     * update the user without validating the data.
+     * update the user without validating the data, only role validation is done.
      * Additionally, if the oldUserDTO is passed, we check if the password was changed. If it was, we
      * need to make sure we encrypt it before updating it.
      */
-    public static boolean updateUserAllFieldsWithoutValidation(UserDTO newUserDTO, UserDTO oldUserDTO) {
+    public static boolean updateUserAllFieldsWithoutDataValidation(UserSession userSession,
+                                                                   UserDTO newUserDTO,
+                                                                   UserDTO oldUserDTO) {
+
+        if (! _isUserHasEditPermission(userSession) || ! _isUserEditingHisOwnData(userSession, oldUserDTO)) {
+            return false;
+        }
+
         if (! newUserDTO.getPassword().equals(oldUserDTO.getPassword())) {
             String newPassword = Utils.encryptPassword(newUserDTO.getPassword());
             newUserDTO.setPassword(newPassword);
@@ -88,15 +128,34 @@ public class DatabaseAPI {
         return UserDAO.updateUserAllFields(newUserDTO);
     }
 
-    public static boolean updateUserFullName(String username, String fullName) {
+    public static boolean updateUserFullName(UserSession userSession, String username, String fullName) {
+        if (! _isUserHasEditPermission(userSession) || ! _isUserEditingHisOwnData(userSession, username)) {
+            return false;
+        }
+
         return UserDAO.updateUserFullName(username, fullName);
     }
 
-    public static boolean updateUserNumber(String username, String number) {
+    public static boolean updateUserNumber(UserSession userSession, String username, String number) {
+        if (! _isUserHasEditPermission(userSession) || ! _isUserEditingHisOwnData(userSession, username)) {
+            return false;
+        }
+
         return UserDAO.updateUserNumber(username, number);
     }
 
-    //User API
+    private static boolean _isUserHasEditPermission(UserSession userSession) {
+        return userSession != null && userSession.isAdmin();
+    }
+
+    private static boolean _isUserEditingHisOwnData(UserSession userSession, UserDTO userDTO) {
+        return userSession != null && userDTO != null && userSession.getUsername().equals(userDTO.getUsername());
+    }
+
+    private static boolean _isUserEditingHisOwnData(UserSession userSession, String username) {
+        return userSession != null && username != null && userSession.getUsername().equals(username);
+    }
+
     public static Set<CallForwardingDTO> loadCallForwardingRecords() {
         return CallForwardingRecordsDAO.loadRecords();
     }
